requirements.txt
Flask==3.0.3
Flask-Mail==0.10.0
Flask-WTF==1.2.1
flask-mysqldb==2.0.0
bcrypt==4.1.3
python-dotenv==1.0.1
WTForms==3.1.2
